+++ pkg/k8s/spec/const.go
@@ -52,8 +52,6 @@ const selectorPrefix = "linstor.linbit.com/"
 // Kubernetes node labels that are an opt-in selector to run piraeus pods when
 // set to "true".
 const (
-	// PiraeusControllerNode label to mark node eligible to run piraeus-controller pods.
-	PiraeusControllerNode = selectorPrefix + "piraeus-controller"
 	// PiraeusSatelliteNode label to mark node eligible to run piraeus-node pods.
 	PiraeusNode = selectorPrefix + "piraeus-node"
 )